summarizer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic>=2.8,<3
python-dotenv==1.0.1
openai>=1.30.0
